[JKImageFlowView] Implement dataSource
This change adds dataSource handling.  Right now, this is a bit
primitive in that in only accepts NSStrings represinging the file
path.  Eventually, it should use something like IKImageBrowserItem.

This change also adds reloadData which uses the dataSource to populate
an NSArray of NSDictionary objects.  The dictionaries contain only
elements for a "url" key.  The NSArray is then set into Image_List
input in the composition.