[JKImageFlowView] Recenter in setFrame
When setting a new frame don't adjust the frames of layers instead
just redraw them and they will snap to the right place.  Also, set the
duration for this animation to 0 so that it happens right away rather
than gliding back to the center.

Do scale the layers based on the frame size when drawing them.