[project] Project Overhaul
Create new projects for MacOS/JKImageFlowView and MacOS/JKImageFlowViewTest.

This organization will allow for separate and parallel iOS projects
that will share code with this project.  Link MacOS/JKImageFlowView with
QuartzCore instead of Qurtz as this will match better with how the iOS
project will be linked.

MacOS/JKImageFlowView is included as a subproject in MacOS/JKImageFlowTest
so that the framework and the test can be built together.

This change also removes the old JKImageFlowView and JKImageFlowViewTest
projects.