[iOS/JKImageFlowView] Initial iOS port
This is a first cut at the port to iOS.  It includes a new
iOS/JKImageFlowView that pulls in the sources from the core
JKImageFlowView.  There is an abstraction of the cocoa differences
between iOS and MacOS.

The MacOS input handling is not appropriate for iOS so it has been
change.