[iOS/JKImageFlowView] Adjust the transforms for iOS
Compute a rootTransform based on the width of UIView.  This is based
on scaling factors that looked good for iPhone and iPad with a linear
ramp between them to handle views of any size.

Adjust the positioning of the title and subtitle so as to display
better on iOS devices.